Update temperature unit used for sending CoE v2 values
